Separate ./js and ./native
