+++ package.json
@@ -36,6 +36,6 @@
     "tape": "^4.9.0"
   },
   "browser": {
-    "./index.js": "./ecurve.js"
+    "./index.js": "./js.js"
   }
 }
